Move createTracer into spa helper
